+++ src/Workspaces/Core/Portable/Shared/Extensions/ITypeSymbolExtensions.cs
@@ -208,28 +208,6 @@ internal static partial class ITypeSymbolExtensions
         return type?.Accept(new UnnamedErrorTypeRemover(compilation));
     }
 
-    [return: NotNullIfNotNull(parameterName: nameof(type))]
-    public static ITypeSymbol? SubstituteTypes<TType1, TType2>(
-        this ITypeSymbol? type,
-        IDictionary<TType1, TType2> mapping,
-        Compilation compilation)
-        where TType1 : ITypeSymbol
-        where TType2 : ITypeSymbol
-    {
-        return type.SubstituteTypes(mapping, new CompilationTypeGenerator(compilation));
-    }
-
-    [return: NotNullIfNotNull(parameterName: nameof(type))]
-    public static ITypeSymbol? SubstituteTypes<TType1, TType2>(
-        this ITypeSymbol? type,
-        IDictionary<TType1, TType2> mapping,
-        ITypeGenerator typeGenerator)
-        where TType1 : ITypeSymbol
-        where TType2 : ITypeSymbol
-    {
-        return type?.Accept(new SubstituteTypesVisitor<TType1, TType2>(mapping, typeGenerator));
-    }
-
     public static bool CanBeEnumerated(this ITypeSymbol type)
     {
         // Type itself is IEnumerable/IEnumerable<SomeType>
