+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/Extensions/ITypeSymbolExtensions.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 
 namespace Microsoft.CodeAnalysis.Shared.Extensions;
 
@@ -22,5 +23,27 @@ internal static partial class ITypeSymbolExtensions
         result ??= [];
         type?.Accept(new CollectTypeParameterSymbolsVisitor(result, onlyMethodTypeParameters: false));
         return result;
+    }
+
+    [return: NotNullIfNotNull(parameterName: nameof(type))]
+    public static ITypeSymbol? SubstituteTypes<TType1, TType2>(
+        this ITypeSymbol? type,
+        IDictionary<TType1, TType2> mapping,
+        Compilation compilation)
+        where TType1 : ITypeSymbol
+        where TType2 : ITypeSymbol
+    {
+        return type.SubstituteTypes(mapping, new CompilationTypeGenerator(compilation));
+    }
+
+    [return: NotNullIfNotNull(parameterName: nameof(type))]
+    public static ITypeSymbol? SubstituteTypes<TType1, TType2>(
+        this ITypeSymbol? type,
+        IDictionary<TType1, TType2> mapping,
+        ITypeGenerator typeGenerator)
+        where TType1 : ITypeSymbol
+        where TType2 : ITypeSymbol
+    {
+        return type?.Accept(new SubstituteTypesVisitor<TType1, TType2>(mapping, typeGenerator));
     }
 }
